+++ user-servcie/src/main/java/com/ssonzm/userservcie/config/security/AuthenticationFilter.java
@@ -1,0 +1,73 @@
+package com.ssonzm.userservcie.config.security;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.ssonzm.userservcie.dto.user.UserRequestDto.UserLoginReqDto;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.core.env.Environment;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+import javax.crypto.SecretKey;
+import java.io.IOException;
+import java.time.Instant;
+import java.util.Base64;
+import java.util.Date;
+
+
+@Slf4j
+public class AuthenticationFilter extends UsernamePasswordAuthenticationFilter {
+
+    private final Environment env;
+
+    public AuthenticationFilter(AuthenticationManager authenticationManager, Environment env) {
+        super(authenticationManager);
+        this.env = env;
+    }
+
+    @Override
+    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException {
+        try {
+            UserLoginReqDto userLoginReqDto = new ObjectMapper().readValue(request.getInputStream(), UserLoginReqDto.class);
+
+            UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(
+                    userLoginReqDto.getEmail(), userLoginReqDto.getPassword()
+            );
+
+            return getAuthenticationManager().authenticate(authenticationToken);
+
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @Override
+    protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response,
+                                            FilterChain chain, Authentication authResult) {
+        PrincipalDetails principalDetails = (PrincipalDetails) authResult.getPrincipal();
+
+        byte[] secretKeyBytes = Base64.getEncoder().encode(env.getProperty("token.secret").getBytes());
+
+        SecretKey secretKey = Keys.hmacShaKeyFor(secretKeyBytes);
+
+        Instant now = Instant.now();
+
+        String userId = String.valueOf(principalDetails.getUser().getId());
+        String token = Jwts.builder()
+                .setSubject(userId)
+                .setExpiration(Date.from(now.plusMillis(Long.parseLong(env.getProperty("token.expiration_time")))))
+                .setIssuedAt(Date.from(now))
+                .signWith(secretKey)
+                .compact();
+
+        response.addHeader("token", token);
+        response.addHeader("userId", userId);
+    }
+}
